Addition of routing, to support the swapping of components and site navigation.
