+++ frontend/package.json
@@ -11,11 +11,13 @@
     "core-js": "^3.6.5",
     "vue": "^2.6.11",
     "vue-image-markup": "^3.1.10",
+    "vue-router": "^3.2.0",
     "vuetify": "^2.2.11"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
     "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-plugin-router": "^4.5.9",
     "@vue/cli-service": "~4.5.0",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
