+++ frontend/src/views/Home.vue
@@ -1,0 +1,16 @@
+<template>
+  <div class="EditorComponent">
+    <EditorComponent/>
+  </div>
+</template>
+
+<script>
+import EditorComponent from '@/components/EditorComponent.vue'
+
+export default {
+  name: 'Home',
+  components: {
+    EditorComponent
+  }
+}
+</script>
